VIRO-1272: Physics Base Integration
- Each VRONode can initalize and contain a strong reference to a VROPhysicsBody.
A VROPhysicsBody contains all the physics proprties upon which forces and
constraints can be applied to. There can be Static, Kinematic and Dynamic
VROPhyscis Bodies. A VROPhysicsBody can be provided a VROPhysicsShape that
effectively represents the collision shape for that body. If none is provided,
one is inferred from the collision box of a given vroNode's geometry.

- Each VROScene can contain a single VROPhysicsWorld. VROPhsyicsBodies are
then added to the physics world to be simulated upon. The VROPhysics world
processes and iterates through the list of active physics bodies within
the render pass in VRORenderer.

- As part of the physics pass, bullet takes in the world transformation of
a node, processes the phsyics, and then outputs it back where we then apply
the updated position and rotations back onto the vroNode.