VIRO-1184 [visibility]: frustum culling
Add basic “far-points” based frustum culling. This
introduces a new computeVisibility() phase into the
render-cycle, and additionally computes an “umbrella”
bounding box for each node. This umbrella box is the
bounds of the node including the bounds of all of its
descendents.

Finally, the sortKeys functions are modified so that
we only update sortKeys for visible nodes; invisible
nodes are skipped, and therefore not rendered.

This commit does not yet take advantage of temporal
cohesion (the TDF frustum)