VIRO-57 [skeletal]: drive skeletal animations
This commit sets up the infrastructure to drive
skeletal animations via the execution API. Previously
we used the execution API for animations to drive
animation groups and chains; it turns out this API
is also idea for key-frame skeletal animations.

This commit:

1. Implements the execution API for skeletal
animations.

2. Enables nodes to contain animations. This
means that FBX can load the animations for a node
into that node directly, simplifying the API for
end-users, and it means skeletal animations can take
advantage of the node hierarchy, starting and stopping
same-named animations all the way down the
hierarchy if needed.