VIRO-1318: AR view, renderer, head-tracker
Created the VROViewAR, which is equivalent to our
VROViewCardboard, except renders an AR view. It
implements the VROView protocol.

This commit also brings back the VROHeadTracker,
which was previously used for the non-Cardboard
Metal renderer. This tracks the phones orientation
using sensors. To enable this simply had to remove
some pre-processor declarations that were excluding
it from compilation in the past, and add support for
head tracking in portrait mode.