VIRO-1317 [AR camera]: camera background
Adds the ability to set a planar texture as a
background. Since the texture is rendered in
screen space, it follows the camera.

Hooked this up to the VROViewAR to get the
camera texture renering as background.

Fixed orientation of the camera texture, and
made it use high quality (previous was very
low).

This commit also changes VROSample to use
the VROViewAR, instead of Cardboard.