VIRO-1317 [AR camera]: support orientation changes
Support phone orientation changes by:

1. Modifying the surface size when the orientation
    changes.
2. Changing the camera orientation.

This required generalizing the setBackground method
of VROScene to accept an arbitrary geometry. That way
the entire background geometry can be configured by the
AR View, instead of half of it being configured by the
scene and the other half by the AR view. So, in turn, the
AR View can change the background size when the
orientation changes.

Finally, note that ideally apps will only have portrait
orientation; everything works great then, even if the phone
is tilted into landscape mode. This commit is to make
landscape-only phones work as well, along with apps
that choose to switch between portrait and landscape,
for whatever reason.