VIRO-1332: derive camera intrinsic matrix
Derive the camera intrinsic projection matrix
using the camera parameters; namely the FOV
and the camera image width and height. When
used as a projection matrix, this ensures the
scale of the virtual world matches the scale of
the real world.

Note that without the correct extrinsic (lookAt)
matrix, objects don’t move perfectly around the
world.