VIRO-1348 [AR]: anchor detection and support
Anchors are real world objects detected by the AR engine.
Once detected, the new AR session delegate can create a
VRONode to correspond with the anchor. By doing so, the
node is kept continually up to date by the AR session. This
node can be used to position virtual objects relative to the
real-world anchor.

For example, if a plane anchor is detected, the can use
its corresponding VRONode to place virtual objects for a
table-top game.

This commit adds support for anchors, adds the new delegate
functions for AR Session, adds planar anchor support, and
adds an example where we add a beating heart to all detected
planar surfaces.