VIRO-1352: lights impact all nodes
Previously lights only impacted their children,
instead of all nodes in the scene. Changed so
that lights now impact all nodes that are within
the attenuationEndDistance of the light.

In each node, _computedLights still represents
each of the lights that influence that node. The
difference is, we now run a pre-pass collectLights
phase to collect all lights in the scene. Those lights
are passed to each child during the updateSortKeys
phase. Only those lights within the attenuation
distance are added to a node’s _computedLights.
The rest of the render pass (calculating the hash
of the lights for sorting, etc.) then runs as normal.