VRO-1371: Fixed Physics Scaling Issues
- Physics shapes were incorrectly constructed as the bounding box we had
previously created for a geometric shape, when trasnformed, was also incorrect.
The fix is to then perform the correct mathematical calculations when
transform a VROBoundingBox.

- Update the geometry's underlying bounding box whenever we update the
geometry. As we were not doing so before, this would also cause physics
shape issues whenever the dimensions of the shape has changed.

- Re-update the physics body whenever we set the orientation of a node
by refreshing it on the onAnimationFinished callback that is triggered
whenever any of these are being set.