VIRO-1426: Compound Shape Scaling and Intertia Bug
- Properly calculate and apply the inverse transform required to
offset/orient the subshapes in reference to it's center of mass
in a compound shape. We were previously incorrectly calculating
positional/rotation offsets, of which was the wrong approach.

- The transforms of Bullet Physics bodies are actually it's center
of mass. This does not necessarily align with Viro's geomtric node
transform, causing misalignment of physics-to-geometric objects.
As such we also now calculate a transform offset to account for
this misalignment for automatically generated physics shapes. This
offset is applied whenever we update the transform of an node as
a result of a physics simulation.

- Also fixed issue where we were not properly calculating the
drag inverse transform for an object when seting the world transform
on a given node.