VIRO-1400 [portals]: separate binding materials from geo
Previously, every time we bound a geometry we would bind
the associated materials properties, even if those properties
has no ties to the geometry. Now that we need a stencil
rendering pass, we want to be able to bind a *single* material
(e.g. one with a quick shader) and rapidly render all the models
just to the stencil buffer, without expensive material switching
logic. To make this happen, we have to separate binding
materials into 3 parts:

1. binding the shader
2. binding material properties that are independent of geometry
3. binding material properties that are geometry-dependent

#1 was already split from #2 and #3. This commit splits apart
#2 and #3, and only binds geometry-independent material
properties when required.