VIRO-1400: change active portal
This commit enables the user to change what
portal is active (e.g. what portal the user is actually
‘in’). This alters the portal tree to make it seem like
the user is inside this other portal, looking out into
the AR world (or whatever other world there is).

This required a little trickery in that we have to
render an ‘exit’ portal to get *back* out to the
previous portal. To do this we use the geometry of
the portal we’ve entered.