VIRO-1400: two-sided portals for clean traversal
Implement two-sided portals to enable clean traversal
from one zone to another. When we’re “near” a portal,
we change that portal to two-sided mode, which means
we only render one of its sides to the stencil buffer. This
has the effect of making the portal two-sided: one side
points to the AR world, the other to the VR world, for
example. By doing this, we remove the ‘flicker’ that
occurs when we pass between worlds.

(This flicker would occur because if we changed active
portals too soon, the portal would suddenly open up a
window to the AR world, even if we hadn’t yet crossed
into the VR world. Or, if we changed too late, we’d clip
the portal, removing its stencil and revealing the AR
world even when we practically in the VR world already).