VIRO-1400: alpha discard stencil writing
Fix bug in portals where the opaque frame of
each portal was written to the stencil buffer; this
caused errors where objects in a different portal
would be rendered atop the frame of the parent
portal.

Fixed by only writing the transparent sections
of the portal frame geometry to the stencil buffer.
To do this, introduced an alpha-discard shader,
which discards opaque fragments.

Also had to introduce textured silhouette
rendering, so that the silhouette shader would
receive the texture (with the alpha in it) to do the
discards correctly.