VIRO-1475: multi-pass renderer architecture
Introduce multi-pass rendering architecture, with the
following key components:

1. VRORenderPass: defines a single rendering pass
that takes a scene and some input textures and writes
to one or more render targets.

2. VRORenderTarget: depth, stencil, or color buffers
that can be written to. Implemented via FBOs in OpenGL.

3. VROChoreographer: sequences the passes and
their targets. May be configured or subclassed for more
advanced rendering techniques.

In implementing this architecture a bit of functionality was
moved, namely:

1. Buffer operations (e.g. color, stencil, depth) were moved
from VRODriver to VRORenderTarget, where they belong
since they are buffer scoped.

2. Scene rendering has moved from VROScene to the new
VROPortalTreeRenderPass. This render pass encapsulates
the entirety of our current rendering technique.

Finally, a special render target for the backbuffer (the
display) was added, so that we can use this the fixed display
in our render pass architecture.