VIRO-1449: record video via FBO and pixel buffers
Record video using the new multi-pass FBO capabilities
and the existing VROVideoCache. The VROVideoCache
is what we previously used to grab frames from a video
and turn them into an OpenGL texture; it turns out this also
works the *other* way around: these special textures are
backed by a CVPixelBuffer, so if we attach one to an FBO
and render to the FBO, we can immediately read the data
back on the CPU via the CVPixelBuffer.

Implemented this system in VROViewAR, replacing the
existing snapshot system.

Note that this also required an additional render to another
FBO because we have to ‘flip’ the texture in order for it
to render right-side-up in video.

Therefore the full render pass is:

1. Render Scene to intermediate FBO
2. Flip and blit intermediate FBO to CVPixelBuffer-backed FBO
    (feed CVPixelBuffer into AVAssetWriter)
3. Blit intermediate FBO to display