Introduce lighting model shader modifier
Create a new shader modifier that enables us
to create a per-light function that adds to illumination
of each fragment.

With this change we can eliminate our lambert,
blinn, and phong shaders and instead implement
each of these as shader modifiers, streamlining
the overall shader model.

Streamlining the shader model is important as we
move toward PBR lighting and shadows.