VIRO-1460, VIRO-1459: Particle Modifiers and Emitter properties
- Created VROParticleModifier that encapsulates the behavior trend of
a particle's property through the usage of interpolated data points. Modifiers
can be created and configured on a Particle Emitter to generate customized
particle behavior and has by default alpha, color, scale, roation, velocity
and acceleration modifiers.

- Implemented the ability to add burst behavior on a particle emitter.
Developers will be able to provide a timeframe or distance at which to
burst n number of particles for c cycles.

- Particle Emitters can now be set to emit N number of particles per
second, or particles per meters. We can also set the duration, loop,
number of particles and lifetime on the particle as well.