VIRO-1498 [shadows]: shadow map generation
New render pass that generates shadow maps on
a depth render target. The depth render target is then
attached to the VRORenderContext so that, during the
base render pass, it can be accessed on a per-light
basis.

Verified by checking the OpenGL snapshots. There are a
number of TODOs in the code related to:

1. Making one shadow map per light.
2. Configuring shadow map properties.