VIRO-1499 [shadows]: per-light shadowing
First steps toward per-light shadowing:

1. We now create a shadow render pass for each
    light that casts shadows. Each of these render passes
    is executed each frame to create a series of depth
    maps.

2. Shadow transform parameters (e.g. the view and
    projection matrices for shadows) are piped into a new
    vertex lighting UBO, allowing for multiple shadow
    coordinates in a single pass.

3. VROLight has been parameterized so the shadow
    render pass uses parameters derived from the
    light.

This commit still only works with a single light. What’s
missing is support for accessing multiple shadow
depth maps in the fragment shader. This will likely be
achieved with a sampler2DArrayShadow.

Additionally, VROLight’s shadow bias and opacity
parameters are not yet respected, and lighting
contribution visibility has to be modified so that it only
impacts the current light.