VIRO-1503 [perspective shadows]: fix bias matrix
We were applying the bias matrix (which converts
shadow coordinates from [-1,1] to [0,1] before
perspective division, which is fine, except we were
not factoring in that the bias matrix ends up
multiplying the w coordinates as well.

Fixed this, which makes perspective shadow maps
work.

Also cleaned up and commented the shadow map
shaders.