Fix material binding crash
This commit fixes the crash caused when we tried
to bind a material without its shader already been bound.

The problem as the logic in VROPortal for rendering
sort keys. Previously we only would bind shaders when
they changed; however, with the new API we have to
call bindShader on a material even if the shader hasn’t
changed, because bindShader now also *selects* what
shader the material should use.

Fixed this and made the system more elegant by:

1. Always calling bindShader whenever the material changes.
    This simplifies the main render-cycle function in VROPortal.

2. Regulate redundant calls to bind shaders via the VRODriver.
    This matches how we regulate redundant calls everywhere
    else.

3. Clean up pglpush and pop so that we can display the
    OpenGL trace by shader, which is much cleaner and
    more informative.