VIRO-1554: could not see original world when pass portal
The issue was that culling wasn’t disabled on the portal
frame, so the portal frame’s transparent region wasn’t
being written to the depth buffer, so the background of the
‘other’ zone was being rendered over this one.

Fixed by force-disabling culling for the entire portal frame.
We have to do this each frame because the portal frame
node tree can change at any time.