VIRO-1578 [bloom]: Gaussian blur + additive blend
Blur the bloom texture then accumulate it into the
HDR buffer. The blur takes places in the Gaussian
render pass, and the accumulation takes place in the
tone mapper. The accumulation happens in the
tone mapper because we want to do this all before
applying the tone mapping and gamma correction
(this way we map the high intensity bloom light
correctly down to RGB)

To enable this flip the driver’s isBloomEnabled method
to true. Disabled for now because of performance,
and because we’re performing Gaussian blur even
when there’s no active material with bloom.