VIRO-1629: portals appearing black
This issue was caused because the material visual’s
swapTexture() function, used when rendering portal
silhouettes, was not properly updating the texture being
used in the material-shader binding.

As a result, silhouettes were using the wrong textures,
and thus writing the wrong portions of the portal
frame to the stencil buffer (the alpha test is dependent
on the texture being used).