VIRO-1569: left eye was rendered twice
VR mode was distorted on iOS because the
viewport settings were causing the left eye to
be rendered twice, instead of the right eye.

This occurred because we were setting the
viewport X on our render targets, which are
eye-size and not screen-size. As a result, when
we rendered the right eye, we weren’t rendering
anything at all. Then, when we blitted to the
display, we would take the existing contents
of the framebuffer (the left eye) and blit those
to the right side of the display.

Fixed so that only the display gets the viewport
X. The render targets always have viewport X
equal to 0.