VIRO-1749, 1740: portal traversal issues
Fix bug where we were traversing into portals
that weren’t in the current portal scene. We now
only check for traversal among the first layer
of portals in the portal tree to fix this.

Also fix crash by checking for null in portal tree,
for cases where the traversal listener is run
before a portal tree is generated.