VIRO-1633: eliminate portal flicker
Eliminate (or at least reduced) portal flicker by
increasing the epsilon multiplier. The higher the
multiplier is, the sooner we switch the active portal.
The goal is to switch *before* the portal gets
clipped, but not too soon that it feels unrealistic.

(if we wait too long to switch, the portal gets clipped
and that results in flicker — we’re in World A, we’re
walking toward a portal displaying World B, as we get
near the portal we pretty much only see World B, but
then the portal gets clipped so we see World A again,
until we traverse the portal at which point we’re back
to World B).