VIRO-1600: use glFramebufferBlit for direct blitting
This helps performance when rendering to video;
instead of drawing a quad, we can directly blit to
the video texture (bypassing fragment shaders).

Unfortunately we can’t directly blit the tone-mapped
HDR output directly to the display framebuffer,
because the display framebuffer is multisampled
(glFramebufferBlit does not support writing to
multisampled FBOs).