Fix missing right eye on Axon
Axon devices change the render target between
rendering eyes. When it came time to bind the
buffer for the right eye, the VRODriver rejected
the change because it thought that FBO was
already bound, when it wasn’t because Axon had
switched it.

Fix this by removing the redundant state change
check. Unfortunately this brings back the spam log
message, but fix this by making the unbind() function
manually unbind the buffer.