Add alpha detection on loaded images
This isn’t used yet, but VROImages now set the
format to RGB8 if there’s no alpha channel. This
format still renders internally as RGBA8, because
RGB8 is not supported with mipmaps (when SRGB
is enabled).

We may use the new hasAlpha() method on
VROTexture for better rendering order computation
in the future.