VIRO-1779: Clicking events being eaten up by invisible VROAnchorViews
There is currently a bug where we incorrectly determine the in-Frustum-View _visbility
of a VRONode, resulting in incorrectly computed opacities, which in turn causes
hidden views to "eat up" events, even though they are not visible. This bug is mitigated
in this fix by only performing hit events on nodes that are both visible and in-Frustum.
The original incorrect visibility bug is tracked at VIRO-1797.