VIRO-1857, VIRO0-1858: Fix particle Flicker and Reload Produces Visual Artifact bugs.
- Improper setting of scale properties to the initial transform of the particle was causing
a flicker to occur, where the particle temporary renders a large scale version itself. The
solution is to ensure we set the initial scale of the transform correctly.

- Particle shaders were also not receiving the right data in their uniform blocks - the
data is being clobbered when a reload occurs. Fix is to bind the buffer base pointers
when bind() is called, similiarly to how we bind data for UBOBones.