VIRO-1321 [arcore]: background texcoord transforms
ARCore has a different way of setting the background
image’s texture coordinates; instead of giving us a
transform like ARKit, it does the transform for us. This
means we have to support directly injecting texcoords
into the VROSurface that’s used as our background.

This commit adds this support, and implements the ARCore
methods for retrieving the transformed texture coordinates.