VIRO-1929 [renderer tests]: create test harness
Decompose VROSample, which was a monolithic
mess of tests, into individual *cross-platform* tests.
These tests are no activated via the
VRORendererTestHarness. Each test implements
VRORendererTest, and returns a SceneController
and point of view.

All of the tests were successfully ported except
for particle test and perhaps ARDraggableNodeTest.

These tests also compile on Android, but do not yet
work because the Android section of VROTestUtil
still needs to be filled out, and the assets need to
be added to Android’s renderertest for this to all
work.

VROSample has been pruned down so it essentially
just calls into the test harness with the ID of the test
it wishes to run. We can attach buttons or a menu later
to make it easier to switch between tests.