VIRO-1850: luminance only Hable tone mapping
Try luminance only Hable tone mapping. This
saturates the colors a bit more than regular Hable.
It’s not perfect and we probably need exposure-
detection based tone mapping to get this just right.
That way we can disable tone mapping entirely
for low-exposure scenes, and only turn it on when
dealing with a high brightness scene. Story being
added to do this.