Allow queuing detached arplanes with ID
Prior to this change, if you had plane1 with id 1 and plane2 with id 2, then
plane2 would *never* be attached to an anchor with id 1 even if plane1 was removed.
With this change, we keep track of plane2 so that if plane1 was to ever be removed
or have its id changed, we will attach plane2 to the anchor with id 1.