VIRO-1983 [Java API]: physics body
Create new Java API for the Physics body; this
is a larger refactor since previously the Node_JNI
had all the physics functions.

Also includes a change whereby we automatically
add a PhysicsBody to the PhysicsWorld whenever
Node::initPhysicsBody is invoked.

To do this we also have to now keep track of the
scene each node belongs to, and ensure the
VROScene gets set whenever a node is attached
to the scene graph.