VIRO-2028: Android Screenshot and Video Recording (Renderer)
- Added VRORenderToTextureDelegate for Android/iOS recorders to hook into for grabbing
rendered frames required for recording / screen capturing.

- Added ViroMediaRecorder.java facilitating Android screen recording and capturing. This
is backed by an egl MediaRecorderSurface.java, that the renderer swaps frame data onto via
the VRORecorderEglSurfaceDisplay render target. This swapping is natively handled by
VROAVRecorderAndroid through the usage of set renderToTextureDelegateAndroid on the
Choreographer to grab the final texture with the rendered scene to be recorded. Note that
the ViroMediaRecorder.java also records audio as well.

- Added a ViroMediaRecorderDelegate.java interface to be implemented by developers for the
notification of sucess or failures when taking a screenshot or recording a video. Encountered
errors during recording would gracefully fail. If the renderer is destroyed during a recording
session, the recording will be forcefully stopped and resources released.

- Also removed any setVRMode implementations for ViroViewARCore.java as they were interfering
with the activity layout orientation.