VIRO-2115: Fixed Crashing Tests
- Removed isReticlePointerFixed API as it was making function calls that needed
to be render-threaded, on the UI thread. This would cause VROThreadRestricted
exceptions to be thrown, causing the crash.

- Replaced the isReticlePointerFixed check in VROGvrLayout to instead query
the controller type. This effectively has the same effect, as only the cardboard
controllers are fixed.

- Because the setup/teardown process of the release test invovles multiple activities,
it can involve multiple Viro renderers being setup / torn down at the same time,
causing crashes (it seemsthat the renderer may not be able to handle this scenario).
The fix is to ensure a single renderer is created and properly destroyed, before setting
up the next test. This is done by pre-emtively calling setup / shutdown activity states
on the renderer, within onResume and onPaused. This is because onResume and onPaused are
gauranteed to be called in order sequentially before the next test.