VIRO-2120 Pausing ARCore while drawing Point Cloud sometimes crashes
We read the point cloud from a FloatBuffer that ARCore gives us, but
when it's suspended, it looks like ARCore flushes the Buffer and
this occasionally happens as we're reading it, so we get a
BufferUnderflowException. In this case, we just check for the Exception
and ignore it, returning no points.