VIRO-1987: Java API AR and VR views
Streamline the API for Java API views. This renames the
views to:

ViroViewGVR, ViroViewOVR, and VivoViewARCore. The
three views all descent from ViroView, which is now a class
instead of an interface. ViroView, furthermore, extends from
FrameLayout. This way each of the subclasses (GVR, OVR,
and ARCore) can implement their platform-specific view
in their own way and make it a *child* of the ViroView.

Javadoc completed on these three classes as well.