VIRO-2183 ARDeclarativeNode crash
In the default constructor of ARDeclarative node, it called all the
way back to Node's default constructor (vs it's dummy constructor).
As a result, Node created a VRONode and invoked initWithNativeRef
which ran ARDeclarativeNode's overridden version that then attempted
to create a native ARNodeDelegate which to reinterpret casts the given
reference to the VRONode as a VROARDeclarativeNode which sometimes
fails (but sometimes succeeds strangely enough). The fix is to simply
call Node's dummy constructor from ARNode's default constructor.