VIRO-2178 Notify ARSessionDelegates of all found anchors when attached
We weren't doing this so this was preventing the declarative manual
users from setting planes that were found in a previous scene because
the current scene didnt know of the existance of previous anchors. Now
when a delegate is attached to the VROARSession, they'll be notified of
all the anchors that have been found up to that point.