VIRO-1953: shadow test intensity errors
There are numerous bugs that have been uncovered
in this test, this fixes one of them: the constructor for
spotlight was not converting radians to degrees. This
is why the shadows only started working after turning
the shadow off and on, because that resets the angles
using a different method (one that does do the
conversion). Note longer term we should change the
native method to take radians as well, as it’s probably
the only place left where we’re still taking degrees.

There are two outstanding bugs in this test:
1. The intensity difference (not sure if this still reproduces)
2. The test itself completely fails if it’s the first test run.
    This probably has something to do with renderer
    initialization.