VIRO-1876: animation transaction crash bugs
These bugs were caused because the FBX loader
would call methods like setPosition and setRotation
off the rendering thread, on nodes that were being
created on a background thread. While this is fine,
since those nodes have not yet been passed off
to the renderer, these methods automatically attempt
to *animate* the setters. This would corrupt the
single-threaded std::maps used in VROTransaction.

Fixed by adding a new ability to check if we’re on the
rendering thread. If we’re off the rendering thread,
animations short-circuit and immediately set to the
final value of the property.