VIRO-2249: transaction crashes in Java
Because in Java we queue transaction operations
in the rendering thread, they ma not *all* be executed
in the scope of a single render cycle. However,
every render cycle we would create an “implicit”
transaction and automatically commit them all.
This was never actually used, but it caused out
transactions to be committed in Java before all the
operations had gone in.

Fixed by removing the implicit transaction code.