VIRO-2293/4/5: crash after unbinding frame
The unbind method in VRODisplayOpenGLAndroidGVR
was unbinding the entire GVR Frame instead of just its
buffer, which then caused crashes as we would try to
render lines (physics tests) or the reticle without a bound
frame.

It’s unclear why we had an unbind() method to begin
with for GVR; it’s not necessary as we don’t have to
unbind the GVR frame in order to render to a different
FBO. This may have been a misunderstanding on how
GVR’s Frame and Swapchain work.