VIRO-1445: update GVR and OVR (Android)
Update the rendering API used by GVR and OVR to
the latest versions. Both of these APIs now have a
‘layering’ system (layers in OVR, buffered viewports in
GVR) that enable us to render the HUD to a separate,
non-async-timewarped, framebuffer. This commit updates
our GVR and OVR to utilize this new functionality so
that our reticle no longer jitters when we turn on async
reprojection.

Because of changs in OVR, the renderEye API was also
changed to take an eyeView matrix instead of an eye
translation matrix. This maximizes platform flexibility,
even though there’s a bit of redundancy between Viro
computing the view matrix in prepareFrame (via the Camera)
*and* the underlying platform computing this same view
matrix. The new render function will be renamed back to
renderEye (from renderEye2) in a future commit), after
updating the iOS GVR API.