Refactor VROViewCardboard in attempt to fix viewport
GVR has a serious bug on iOS where if an app starts
in portrait mode, it will never render the viewports
correctly. When the user fits the device into the headset,
the viewports take up only half the screen because GVR
somewhere cached the portrait screen size (or something).

This bug can only be worked around by using a
ViewController that forces the orientation to landscape
in viewWillAppear. Doing this is viewDidAppear is too late.
Unfortunately, doing this in viewWillAppear can cause
layout corruption unless the ViewController is pushed
into a NavigationController.

This commit makes our system more closely resemble
Google’s GVRKit wrapper. Unfortunately this didn’t fix
anything but it does make the code cleaner, keeping the
C++ GVR classes restricted to the scene renderer.