Draw pencil during render pass not after
The pencil was being drawn after the main
render pass was complete, when it should be
drawn as part of the render pass. Furthermore,
we were "drawing" the shadow frusta once
per eye, resulting in a doubling up of the lines.
Instead draw the frustra once, during the
shadow pass (which is run once per frame
instead of per eye).