UBOs: use static binding points
Binding points do not need to be created at
runtime, and they do not need to be re-linked
to the shader each time a UBO is bound.
Instead statically specify the binding points
and bind them once to their shader index after
a shader is compiled.

This seems to mitigate the mysterious Adreno
BufferObjectDisableReorderCheck crash.