VIRO-2286: OVR skeletal animation freeze bug
FBX models with animations were freezing on
OVR unless they were loaded synchronously, or
unless the UBO for the skeletal animation was
completely disabled.

It turns out this was a result of passing in uninitialized
data for the bones UBO. If we initialize the data to
identity, then everything works per normal. It's
unclear why this only impacted OVR. However, it
is clear why loading synchronously mitigated
the issue: when loaded synchronously, the probability
of the UBO being rendered with uninitialized data
went down.