VIRO-2364: React Native AR Gaze Interaction
Move the processing of gaze events from the base input controller into
the corresponding platform specific controllers (note that the logic is still
within the base input controller). This enables additional flexibility
for controllers that want to override gazing functionality, such as it is
in the case for AR input controllers. For AR, gaze / hovering callbacks are
done so from the camera's point of view.