VIRO-2414: fix AR hit test events
These were broken because the projection
and view matrices for the HUD (instead of
the eye) were being used in the AR hit test
controller. The root of the bug is that nothing
should hold onto a VRORenderContext, since
that's a transitory object used only in the midst
of a render cycle.

Fix this by disallowing access to the
VRORenderContext, and replacing all such
access with instead direct access to the last
used camera via renderer->getCamera, or
by pushing the modelview and projection
matrices themselves onto the objects that
need them.