VIRO-2426: Crash in Audio Memory Leak Test
Because the PlatformUtils can be destroyed before sound objects are destroyed, this causes
a null point crash when the sound objects attempts to access PlatformUtils as a part of their
cleanup / destruction code paths. As such, the patch here is to apply a null check before
attempting to clean up for our sound objects. I've also added VIRO-2441 to ensure that
we properly clean up our sound and media files in similiar situations.