VIRO-2446 Screen recording not working on iOS
When we record the screen, we capture the video (from the renderer)
and the audio separately. However, due to the latest update to GVR
Audio, for some reason the gvr audio API keeps setting the AVAudioSession
category to SoloAmbient whereas we need it to be either in Record or
PlayAndRecord state in order to record audio. As a result, we were
capturing very short snippets of audio and this was causing our
videos to be very short (because we were using audio length).

Since this seems to be a bug with GVR Audio and the fact that we aren't
really using GVR Audio (due to a previous bug that we haven't yet come
back to), the workaround now is to simply disable GVR Audio and avoid using
the GVR AudioApi for now.

Also reverting a previous change that addressed another symptom of the
GVR Audio bug.