VIRO-2499: shadow map crash on iPhone 6
Shadows were causing a crash on iPhone 6
because in the Hello World AR sample we
were using a large shadow map size (2048)
which is not supported by that device.

Add logic in the renderer to halve the size of
the shadow maps until we find a supported
size. This also replaces hard assert failures
when creating FBOs with soft failures so that
we can recover from them.