VIRO-2350: support old Adreno GPUs
Adreno 330 and older have poor support for GLSL
in that:

1. EGL videos end up with a purple tint
2. Shader compilation with lights fail

These both seem to be tied to the use of structs in the
shader code. Rather than reduce *all* of our shaders
to a primitive format without structs to support Adreno,
we instead introduce the ability to identify what GPU
we're using. If we detect Adreno 330 or earlier, we do
the following:

1. Force the device to use constant lighting.
2. Eliminate the use of structs in the remaining shaders.

We may add more to this list as more Adreno defects
are discovered.